+++ src/Constants.php
@@ -5,5 +5,5 @@ namespace Microsoft\Kiota\Http;
 final class Constants
 {
     /** @var string The current version for this Library */
-    public const KIOTA_HTTP_CLIENT_VERSION = '1.0.0';
+    public const KIOTA_HTTP_CLIENT_VERSION = '1.1.0';
 }